+++ packages/terra-time-input/src/terra-dev-site/doc/example/TimeInputNoDefaultTime.jsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import TimeInput from 'terra-time-input/lib/TimeInput';
+import TimeInput from 'terra-time-input';
 import classNames from 'classnames/bind';
 import styles from './TimeInputDocCommon.module.scss';