+++ packages/terra-application-navigation/src/terra-dev-site/test/application-navigation/HeroDrawer.test.jsx
@@ -6,7 +6,7 @@ import Placeholder from 'terra-doc-template/lib/Placeholder';
 
 const hero = <Placeholder title="Test Hero" variant="light" />;
 
-// TODO: remove terra-application after it is incorporated into dev-site
+// TODO: remove terra-application after it is incorporated into dev-site for themes or themes are co-located
 const HeroDrawerTest = () => (
   <ApplicationBase locale="en">
     <ApplicationNavigation hero={hero} />