+++ packages/terra-application-navigation/src/terra-dev-site/test/application-navigation/OnlyCustomUtilityItems.test.jsx
@@ -45,7 +45,7 @@ const handleItemSelection = (key, metaData) => {
 };
 /* eslint-enable no-alert */
 
-// TODO: remove terra-application after it is incorporated into dev-site
+// TODO: remove terra-application after it is incorporated into dev-site for themes or themes are co-located
 const OnlyCustomUtilityItemsTest = () => (
   <ApplicationBase locale="en">
     <ApplicationNavigation