+++ packages/terra-application-navigation/src/terra-dev-site/test/application-navigation/Tabs.test.jsx
@@ -66,7 +66,7 @@ const handleItemSelection = (key, metaData) => {
 };
 /* eslint-enable no-alert */
 
-// TODO: remove terra-application after it is incorporated into dev-site
+// TODO: remove terra-application after it is incorporated into dev-site for themes or themes are co-located
 const TabsTest = () => (
   <ApplicationBase locale="en">
     <ApplicationNavigation