+++ packages/terra-application-navigation/src/terra-dev-site/test/application-navigation/User.test.jsx
@@ -23,7 +23,7 @@ const userConfigSrc = {
   imageSrc: 'test-src',
 };
 
-// TODO: remove terra-application after it is incorporated into dev-site
+// TODO: remove terra-application after it is incorporated into dev-site for themes or themes are co-located
 const UserTest = () => (
   <ApplicationBase locale="en">
     <FixedHeight><ApplicationNavigation userConfig={userConfig} /></FixedHeight>