+++ packages/terra-date-picker/src/react-datepicker/index.jsx
@@ -12,15 +12,10 @@ import PopupContainer from './PopupContainer';
 import {
   newDate,
   now,
-  cloneDate,
   isMoment,
   isDate,
   isBefore,
   isAfter,
-  setTime,
-  getSecond,
-  getMinute,
-  getHour,
   getMonth,
   addDays,
   addMonths,
@@ -322,30 +317,6 @@ class DatePicker extends React.Component {
      * Prop to close calendar dropdown after date is selected.
      */
     shouldCloseOnSelect: PropTypes.bool,
-    /**
-     * Prop to show selected date and time.
-     */
-    showTimeSelect: PropTypes.bool,
-    /**
-     * Format of the selected time.
-     */
-    timeFormat: PropTypes.string,
-    /**
-     * Interval between 2 consecutive times on the time picker.
-     */
-    timeIntervals: PropTypes.number,
-    /**
-     * Minimum value of time that can be selected .
-     */
-    minTime: PropTypes.object,
-    /**
-     * Maximum value of time that can be selected .
-     */
-    maxTime: PropTypes.object,
-    /**
-     * Array to store values of time that are disabled to pick .
-     */
-    excludeTimes: PropTypes.array
   }
 
   static get defaultProps () {
@@ -367,8 +338,6 @@ class DatePicker extends React.Component {
       monthsShown: 1,
       withPortal: false,
       shouldCloseOnSelect: true,
-      showTimeSelect: false,
-      timeIntervals: 30
     }
   }
 
@@ -389,7 +358,15 @@ class DatePicker extends React.Component {
     document.addEventListener('keydown', this.handleKeydown);
   }
 
-  componentDidUpdate() {
+  componentDidUpdate(prevProps) {
+    const currentMonth = prevProps.selected && getMonth(prevProps.selected)
+    const nextMonth = this.props.selected && getMonth(this.props.selected)
+    if (prevProps.inline && currentMonth !== nextMonth) {
+      this.setPreSelection(this.props.selected)
+    }
+    if (prevProps.highlightDates !== this.props.highlightDates) {
+      this.setState({'highlightDates': getHightLightDaysMap(this.props.highlightDates)})
+    }
     // Shift focus into popup date-picker if it exists
     if (this.datePickerPopupContainer.current) {
       this.datePickerPopupContainer.current.focus();
@@ -398,17 +375,6 @@ class DatePicker extends React.Component {
     // Shift focus into overlay date-picker if it exists
     if (this.datePickerOverlayContainer.current) {
       this.datePickerOverlayContainer.current.focus();
-    }
-  }
-
-  componentWillReceiveProps (nextProps) {
-    const currentMonth = this.props.selected && getMonth(this.props.selected)
-    const nextMonth = nextProps.selected && getMonth(nextProps.selected)
-    if (this.props.inline && currentMonth !== nextMonth) {
-      this.setPreSelection(nextProps.selected)
-    }
-    if (this.props.highlightDates !== nextProps.highlightDates) {
-      this.setState({'highlightDates': getHightLightDaysMap(nextProps.highlightDates)})
     }
   }
 
@@ -539,7 +505,7 @@ class DatePicker extends React.Component {
       }
     )
     this.setSelected(date, event)
-    if (!this.props.shouldCloseOnSelect || this.props.showTimeSelect) {
+    if (!this.props.shouldCloseOnSelect) {
       this.setPreSelection(date)
     } else if (!this.props.inline) {
       this.setOpen(false)
@@ -557,13 +523,6 @@ class DatePicker extends React.Component {
 
     if (!isSameDay(this.props.selected, changedDate) || this.props.allowSameDay) {
       if (changedDate !== null) {
-        if (this.props.selected) {
-          changedDate = setTime(newDate(changedDate), {
-            hour: getHour(this.props.selected),
-            minute: getMinute(this.props.selected),
-            second: getSecond(this.props.selected)
-          })
-        }
         this.setState({
           preSelection: changedDate
         })
@@ -601,21 +560,6 @@ class DatePicker extends React.Component {
     setTimeout(() => {
       this.ariaLiveStatus = '';
     }, 1000);
-  }
-
-  handleTimeChange = (time) => {
-    const selected = (this.props.selected) ? this.props.selected : this.getPreSelection()
-    let changedDate = setTime(cloneDate(selected), {
-      hour: getHour(time),
-      minute: getMinute(time)
-    })
-
-    this.setState({
-      preSelection: changedDate
-    })
-
-    this.props.onChange(changedDate)
-    this.setOpen(false)
   }
 
   onInputClick = () => {
@@ -726,7 +670,6 @@ class DatePicker extends React.Component {
         showMonthDropdown={this.props.showMonthDropdown}
         showWeekNumbers={this.props.showWeekNumbers}
         showYearDropdown={this.props.showYearDropdown}
-        withPortal={this.props.withPortal}
         forceShowMonthNavigation={this.props.forceShowMonthNavigation}
         scrollableYearDropdown={this.props.scrollableYearDropdown}
         todayButton={this.props.todayButton}
@@ -738,13 +681,6 @@ class DatePicker extends React.Component {
         onDropdownFocus={this.handleDropdownFocus}
         onMonthChange={this.props.onMonthChange}
         dayClassName={this.props.dayClassName}
-        showTimeSelect={this.props.showTimeSelect}
-        onTimeChange={this.handleTimeChange}
-        timeFormat={this.props.timeFormat}
-        timeIntervals={this.props.timeIntervals}
-        minTime={this.props.minTime}
-        maxTime={this.props.maxTime}
-        excludeTimes={this.props.excludeTimes}
         className={this.props.calendarClassName}
         yearDropdownItemNumber={this.props.yearDropdownItemNumber}
       >
@@ -783,7 +719,6 @@ class DatePicker extends React.Component {
         showMonthDropdown={this.props.showMonthDropdown}
         showWeekNumbers={this.props.showWeekNumbers}
         showYearDropdown={this.props.showYearDropdown}
-        withPortal={this.props.withPortal}
         forceShowMonthNavigation={this.props.forceShowMonthNavigation}
         scrollableYearDropdown={this.props.scrollableYearDropdown}
         todayButton={this.props.todayButton}
@@ -794,13 +729,6 @@ class DatePicker extends React.Component {
         onDropdownFocus={this.handleDropdownFocus}
         onMonthChange={this.props.onMonthChange}
         dayClassName={this.props.dayClassName}
-        showTimeSelect={this.props.showTimeSelect}
-        onTimeChange={this.handleTimeChange}
-        timeFormat={this.props.timeFormat}
-        timeIntervals={this.props.timeIntervals}
-        minTime={this.props.minTime}
-        maxTime={this.props.maxTime}
-        excludeTimes={this.props.excludeTimes}
         className={this.props.calendarClassName}
         yearDropdownItemNumber={this.props.yearDropdownItemNumber}
       >