+++ packages/terra-date-picker/src/terra-dev-site/doc/example/DatePickerDefault.jsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import PropTypes from 'prop-types';
 import Field from 'terra-form-field';
-import DatePicker from 'terra-date-picker/lib/DatePicker';
+import DatePicker from 'terra-date-picker';
 import classNames from 'classnames/bind';
 import styles from './DatePickerExampleCommon.module.scss';